renamed constants.lambda-shared.ts to constants.ts
