+++ src/backend/ingestion/ingestion.lambda.ts
@@ -5,7 +5,7 @@ import { validateAssembly } from '@jsii/spec';
 import type { Context, SQSEvent } from 'aws-lambda';
 import { extract } from 'tar-stream';
 import * as aws from '../shared/aws.lambda-shared';
-import * as constants from '../shared/constants.lambda-shared';
+import * as constants from '../shared/constants';
 import { requireEnv } from '../shared/env.lambda-shared';
 import { IngestionInput } from '../shared/ingestion-input.lambda-shared';
 import { integrity } from '../shared/integrity.lambda-shared';
