+++ src/backend/inventory/constants.ts
